+++ signin.html
@@ -41,7 +41,7 @@
             <i class="fa fa-eye fa-lg"></i>
             <div id="password-errorText" class="errortext error"></div>
           </div>
-          <button id="btn" class="signin">로그인</button>
+          <button type="submit" id="btn" class="signin">로그인</button>
         </form>
         <div class="social">
           <div class="social-text">소셜 로그인</div>
@@ -56,7 +56,9 @@
         </div>
       </div>
     </main>
+    <script type="module" src="email.js"></script>
+    <script type="module" src="password.js"></script>
     <script type="module" src="signin.js"></script>
-    <script typer="module" src="passwordToggle.js"></script>
+    <script src="passwordToggle.js"></script>
   </body>
 </html>
