+++ signup.html
@@ -70,6 +70,6 @@
       </div>
     </section>
     <script src="signup.js"></script>
-    <script typer="module" src="passwordToggle.js"></script>
+    <script src="passwordToggle.js"></script>
   </body>
 </html>
